Add GeneticAlgorithm class

+++ include/params.hpp
@@ -1,13 +1,15 @@
 #ifndef PARAMS_HPP_
 #define PARAMS_HPP_
 
+#include <cstdlib>
+
 
 class Params
 {
 public:
     static int iterations_number;
 
-    static int population_size;
+    static size_t population_size;
     // crossover probability
     static double pk;
     // mutation probability
