+++ src/params.cpp
@@ -1,8 +1,10 @@
 #include "params.hpp"
 
+#include <cstdlib>
+
 
 int Params::iterations_number = 1000;
 
-int Params::population_size = 250;
+size_t Params::population_size = 250;
 double Params::pk = 0.7;
 double Params::pm = 0.1;
